+++ console/views.py
@@ -1,11 +1,15 @@
 import logging
-import requests
 
+from datetime import datetime
 from flask import render_template
 from flask import request
 import flask_security
+import requests
+from sqlalchemy import func
+from sqlalchemy.exc import DataError, SQLAlchemyError
 from structlog import wrap_logger
 
+from console.database import db, SurveyResponse
 from console import app
 from console import settings
 from console.helpers.exceptions import ClientError, ServiceError
@@ -20,9 +24,14 @@ def home():
     return "stuff"
 
 
-def send_data(url, data):
+def send_data(url, data=None, request_type="POST"):
     try:
-        r = requests.post(url, data)
+        if request_type == "POST":
+            logger.info("Posting data to " + url)
+            r = requests.post(url, data)
+        else:
+            logger.info("Sending GET request to " + url)
+            r = requests.get(url)
     except requests.exceptions.ConnectionError as e:
         logger.error('Could not connect to ' + url, response="Connection error")
         raise e
@@ -50,7 +59,7 @@ def decrypt():
 
         try:
             logger.info("Posting data to sdx-decrypt", user=flask_security.core.current_user.email)
-            decrypt_response = send_data(url, data)
+            decrypt_response = send_data(url, data, "POST")
         except ClientError:
             error = 'Client error'
         except ServiceError:
@@ -65,3 +74,51 @@ def decrypt():
 
     else:
         return render_template('decrypt.html')
+
+
+def get_filtered_responses(tx_id, ru_ref, survey_id, datetime_earliest, datetime_latest):
+    try:
+        q = db.session.query(SurveyResponse)
+        if tx_id != '':
+            q = q.filter(SurveyResponse.tx_id == tx_id)
+        if ru_ref != '':
+            q = q.filter(SurveyResponse.data["metadata"]["ru_ref"].astext == ru_ref)
+        if survey_id != '':
+            q = q.filter(SurveyResponse.data["survey_id"].astext == survey_id)
+        dt_column = func.date(SurveyResponse.data["submitted_at"].astext)
+        if datetime_earliest:
+            year = int(datetime_earliest[:4])
+            month = int(datetime_earliest[5:7])
+            day = int(datetime_earliest[8:10])
+            hour = int(datetime_earliest[11:13])
+            minute = int(datetime_earliest[14])
+            q = q.filter(dt_column > datetime(year, month, day, hour, minute))
+        if datetime_latest:
+            year = int(datetime_latest[:4])
+            month = int(datetime_latest[5:7])
+            day = int(datetime_latest[8:10])
+            hour = int(datetime_latest[11:13])
+            minute = int(datetime_latest[14])
+            q = q.filter(dt_column < datetime(year, month, day, hour, minute))
+        filtered_data = q.all()
+    except DataError as e:
+        logger.error("Invalid search term", error=e)
+        return []
+    except SQLAlchemyError as e:
+        logger.error("Database error", error=e)
+        return []
+
+    return filtered_data
+
+
+@app.route('/store', methods=['GET'])
+@flask_security.roles_required('SDX-Developer')
+def store():
+    tx_id = request.args.get('tx_id', type=str, default='')
+    ru_ref = request.args.get('ru_ref', type=str, default='')
+    survey_id = request.args.get('survey_id', type=str, default='')
+    datetime_earliest = request.args.get('datetime_earliest', type=str, default='')
+    datetime_latest = request.args.get('datetime_latest', type=str, default='')
+
+    store_data = get_filtered_responses(tx_id, ru_ref, survey_id, datetime_earliest, datetime_latest)
+    return render_template('store.html', data=store_data)